+++ data/src/main/kotlin/dev/esbi/mizan/data/local/entity/DashboardSummaryEntity.kt
@@ -2,14 +2,15 @@ package dev.esbi.mizan.data.local.entity
 
 import androidx.room.Entity
 import androidx.room.PrimaryKey
+import java.math.BigDecimal
 
 @Entity(tableName = "dashboard_summary")
 data class DashboardSummaryEntity(
     @PrimaryKey val id: Int = 1,
-    val totalBalance: Double,
-    val monthlyExpenses: Double,
-    val monthlySavings: Double,
-    val budgetLimit: Double,
-    val budgetPercentageUsed: Double,
+    val totalBalance: BigDecimal,
+    val monthlyExpenses: BigDecimal,
+    val monthlySavings: BigDecimal,
+    val budgetLimit: BigDecimal,
+    val budgetPercentageUsed: BigDecimal,
     val lastUpdated: Long = System.currentTimeMillis()
 )
