+++ data/src/main/kotlin/dev/esbi/mizan/data/local/entity/FinancialProjectionEntity.kt
@@ -3,10 +3,12 @@ package dev.esbi.mizan.data.local.entity
 import androidx.room.Entity
 import androidx.room.PrimaryKey
 
+import java.math.BigDecimal
+
 @Entity(tableName = "financial_projections")
 data class FinancialProjectionEntity(
     @PrimaryKey val year: String,
-    val conservative: Double,
-    val realistic: Double,
-    val optimistic: Double
+    val conservative: BigDecimal,
+    val realistic: BigDecimal,
+    val optimistic: BigDecimal
 )
