+++ data/src/main/kotlin/dev/esbi/mizan/data/local/entity/SubscriptionEntity.kt
@@ -3,11 +3,13 @@ package dev.esbi.mizan.data.local.entity
 import androidx.room.Entity
 import androidx.room.PrimaryKey
 
+import java.math.BigDecimal
+
 @Entity(tableName = "subscriptions")
 data class SubscriptionEntity(
     @PrimaryKey(autoGenerate = true) val id: Long = 0,
     val name: String,
-    val amount: Double,
+    val amount: BigDecimal,
     val billingCycle: String,
     val nextRenewalDate: Long,
     val icon: String,
