+++ data/src/main/kotlin/dev/esbi/mizan/data/repository/CurrencyRepositoryImpl.kt
@@ -29,7 +29,7 @@ class CurrencyRepositoryImpl @Inject constructor(
     }
 
     override suspend fun updateRate(code: String, rateToBase: Double) {
-        currencyDao.updateRate(code, rateToBase)
+        currencyDao.updateRate(code, BigDecimal.valueOf(rateToBase))
     }
 
     override suspend fun getCurrencyByCode(code: String): Currency? {
